+++ include/ossc/oss_extra.h
@@ -46,8 +46,8 @@ typedef struct extra_curl_request_param_s extra_curl_request_param_t;
 struct extra_buffer_s {
 	char *unmovable_buffer_ptr; /**<记录主线程中分配的缓冲区首地址，用于在子线程中释放该空间 */
 	char *ptr; /**< 缓冲区首指针，发送数据时该指针需要向后移动 */
-	size_t left; /** 缓冲区剩余大小 */
-	size_t allocated; /** 缓冲区总大小 */
+	unsigned int left; /** 缓冲区剩余大小 */
+	unsigned int allocated; /** 缓冲区总大小 */
 	unsigned short code; /**返回码 */
 };
