+++ include/ossc/util/oss_ttxml.h
@@ -58,6 +58,6 @@ XmlNode * xml_find(XmlNode *xml, const char *name);
  *        XmlNode *xml_parse(const char *buffer),
  * to fit our need.
  * */
-XmlNode *xml_load_buffer(const char *buffer, size_t buffer_len);
+XmlNode *xml_load_buffer(const char *buffer, unsigned int buffer_len);
 
 #endif /* CCAN_TTXML_H */
